feat: New site generates a random, secure password (redux) (#194)
* build: Add `@automattic/generate-password` dependency

Enable simple generation of random, secure passwords.

* build: Add `patch-package` dependency

Enable patching `@automattic/generate-password` package for use in a
Node.js environment.

* build: Patch `@automattic/generate-password` for Node.js compatibility

The `window` global is unavailable in the Node.js runtime, which is used
for Electron's main process.

* feat: Generate random, secure password for new sites

Mitigate vulnerability from the default admin password.

* refactor: Fix function name typo

* test: Repair `content-tab-settings` tests

Currently, the decrypted promise does not result in user-facing changes.
So, we must await the promise resolution. In future changes, we can
await the user-facing changes.

* fix: Account for legacy sites lacking a stored password

Without a migration, pre-existing sites will not have a stored password.
Therefore, we must account for the lack of a password by providing the
previous default value.

* fix: Disable copy password button during asynchronous descryption

Expand `CopyTextButton` to manage this state to simplify scenarios where
the text value is asynchronously fetched prior to passing the value to
the `CopyTextButton`. E.g., decrypting values for copying.

* refactor: Remove errant import

* test: Expand `ContentTabSettings` tests

* test: Add `passwords` lib tests

* refactor: Align wp-now config and CLI options

The `adminPassword` and `siteTitle` options were implemented
simultaneously in separate branches. This change aligns the
implementations for consistency.

* fix: Provide missing `adminPassword` and `siteTitle` arguments to CLI

Finish the incomplete work of expanding the CLI arguments to include
those required by Studio.

* feat: Remove admin password encryption

Simplify the implementation for storing admin passwords for local sites.
The `safeStorage` API requires user-granted OS permissions, which may be
an unnecessary burden for storing these local site passwords.

* test: Remove lingering unused import