fix: Consistent, predictable focus management (#192)
* fix: Markup describes top-level document structure

Simplify comprehension of top-level structure for users relying upon
assistive technologies.

* fix: Remove in-actionable button

Mitigate user confusion from an in-actionable button that is focusable
via keyboards and screen readers.

* feat: Overview settings screen headings describe content structure

Simplify comprehension of content structure for users relying upon
assistive technologies.

* fix: Local path input is focusable

Without a "button" role, the input is not focusable by keyboards.

* fix: Enable button focus outline, refactor secondary variant

The lack of a focus styles can disorient users relying upon keyboard
navigation. The usage of the `!important` modifier caused specificity
collisions, so it was replaced with arbitrary Tailwind variants.

* fix: Replace custom `ButtonSection` styles with secondary variant

Improve consistency in styles.

* fix: Reinstate "Add site" button focus

The lack of focus styles can disorient users relying upon keyboard
navigation.

* fix: Site action button displays focus styles

The lack of focus styles can disorient users relying upon keyboard
navigation.

* fix: Avoid interpolation in Tailwind class names

This appears to break Tailwind compilation.

* feat: Leverage `isDestructive` for delete site button

* feat: Reinstate text input focus style

The lack of focus styles can disorient users relying upon keyboard
navigation.

* feat: Leverage "link" variant for buttons

Reduce need to override button styles.

* feat: Add local path focus style

The lack of focus styles can disorient users relying upon keyboard
navigation.

* feat: Add stop all sites button focus styles

The lack of focus styles can disorient users relying upon keyboard
navigation.

* fix: Improve focus management for dropdown menus

Revert addition of a custom dropdown component that was added to manage
closing dropdown menus when focus is lost.

https://github.com/Automattic/local-environment/pull/68

This custom logic is not necessary. The original issues is that React's
StrictMode runs all components twice, which results in the clean up
function clearing the result of the detached, sibling callback-ref-based
setup function. This means the focus is never placed on the first
focusable item when the dropdown menu opens, which means the dropdown
menu does not close as it never loses focus.

However, since StrictMode is only run in development environments, this
issue is not present in production builds. Therefore, this custom
dropdown menu component is not required.

- https://react.dev/reference/react/StrictMode
- https://react.dev/reference/react/useEffect#my-cleanup-logic-runs-even-though-my-component-didnt-unmount
- https://github.com/WordPress/gutenberg/blob/c50e0c408e7740c6a8a817c985831670d300f030/packages/compose/src/hooks/use-focus-on-mount/index.js#L53-L59
- https://github.com/WordPress/gutenberg/blob/c50e0c408e7740c6a8a817c985831670d300f030/packages/compose/src/hooks/use-focus-on-mount/index.js#L71-L77

* fix: Add site modal returns focus to modal toggle button

Using the `autoFocus` prop for the text control disabled the modal focus
management. This resulted in the focus no longer being returned to the
button toggling the modal visibility as expected. Leveraging the
`focusOnMount="firstContentElement"` ensures the modal focuses the first
focusable element within the modal content, while preserving the ability
to return focus to the modal toggle button.

It is worth noting that focusing the input field does not work in the
development environment due to the usage of StrictMode, which was
referenced in the previous commit as well.

React's StrictMode runs all components twice, which results in the clean
up function clearing the result of the detached, sibling
callback-ref-based setup function. This means the focus is never placed
on the first focusable item when the dropdown menu opens, which means
the dropdown menu does not close as it never loses focus.

However, since StrictMode is only run in development environments, this
issue is not present in production builds. Therefore, this custom
dropdown menu component is not required.

- https://react.dev/reference/react/StrictMode
- https://react.dev/reference/react/useEffect#my-cleanup-logic-runs-even-though-my-component-didnt-unmount
- https://github.com/WordPress/gutenberg/blob/c50e0c408e7740c6a8a817c985831670d300f030/packages/compose/src/hooks/use-focus-on-mount/index.js#L53-L59
- https://github.com/WordPress/gutenberg/blob/c50e0c408e7740c6a8a817c985831670d300f030/packages/compose/src/hooks/use-focus-on-mount/index.js#L71-L77

* refactor: Remove duplicative class name

* feat: Add site menu focus styles

The lack of focus styles can disorient users relying upon keyboard
navigation.

* feat: Add focus styles to preview link button

The lack of focus styles can disorient users relying upon keyboard
navigation.

* feat: `CopyTextButton` leverages WordPress button focus styles

The previous design displayed the default Chromium focus styles, which
did not match the rest of the product UI.

* feat: Open local path leverages WordPress button focus styles

The previous design displayed the default Chromium focus styles, which
did not match the rest of the product UI.

* feat: Improve new snapshots UI accessibility

Leverage the Button component to inherit consistent styles. Hide
decorative images from screen readers. Update preview site image label
to more accurately describe the intent --- a preview of the site, not
the active theme name.