fix: Prevent crashes when starting servers (#223)
* feat: Catch errors when starting servers

Mitigate a disruptive UX by displaying an alert rather than crashing the
app when starting the server fails.

* fix: Disable thumbnail update if server start fails

Updating the thumbnail throws an error if the server is not running.

* docs: Update comment phrasing

* refactor: Remove now unnecessary early return

This was originally placed to avoid thumbnail update failures, but was
misguided and is now handled in `SiteServer.start`.

* test: Assert server start outcomes

* feat: Server start failure message empowers users to fix the issue

Provide users a way to succeed on their own before contacting support.

* fix: IPC `startServer` catches errors

Refactor to relying upon a `try`/`catch` block instead of the
`SiteServer.running` attribute, as this broadens the error handling
coverage that can occur when starting a server with `wp-now`. E.g.,
`getWpNowConfig` can throw errors, which were not caught by the previous
implementation.

* fix: Avoid collapsible white space in translation strings

These characters are incompatible with the i18n tooling.

* refactor: Revert unnecessary changes

This were accidental and have no impact.

* feat: Adjust error message for clarity

Reference the "local path" phrase as that is also used in the site's
settings panel.