feat: Improve form accessibility (#259)
* fix: Prevent invalid site creation during on-boarding

We must prevent site creation if the site's local path is not a valid
directory --- e.g. it is not empty.

* feat: Local path communicates validation status

Expand validation communication so that it is clear for both sighted
users and those relying upon assistive technologies.

* feat: Use warning icon for site creation error message

Communicate the error message is not a tip, but a severe warning that
blocks site creation.

* fix: Apply missing preview site string translation

The `sprintf` function does not translate, it only formats.

* feat: Announce site creation

Improve UX for users relying upon assistive technologies.

* feat: Announce on-boarding site creation

Improve UX for users relying upon assistive technologies.

* feat: Announce site start and stop events

Improve UX for users relying upon assistive technologies.

* feat: Reorder start/stop site a11y label terms

Improve consistency with other accessibility labels and announcements in
the app.

* feat: Add combined Studio description accessibility label

Navigating these terms individually as the very first elements when
launch the app feels a bit disorienting. A singular label communicates
the abstract description in one step, allowing the user to quickly move
forward with the next step of creating a site.

* fix: Reinstate tip icon for informative messages

This logic is used for both error and informational messages, so we need
support for both icons.

* fix: Display "not-allowed" cursor for disabled local path input

The child elements do not need cursor styles, only the top-level button.
The additional styles resulted in a unexpected pointer cursor when the
local path input was disabled.

* fix: Declare missing Hook dependencies

Avoid stale Hooks from missing dependencies.