fix: Catch errors from servers leveraging non-`wordpress` modes (#251)
* fix: Enforce `wp-now` `wordpress` mode

Avoid unexpected mutations (e.g., creating a `wp-config.php` file) to
the WordPress files bundled with Studio. These mutations resulted in
breakages in other Studio logic, and the WordPress server would fail to
connect to the database.

At the present, there is no intention to support non-`wordpress`-mode
sites in Studio.

* refactor: Simplify enforce `wordpress` mode logic

Given this is a Studio-only constraint, there is no need to integrate
this option into `getWpNowConfig`.

* fix: Reverse unintended bundled WordPress installation mutations

Prior to enforcing Playground's WordPress mode, it was possible to run
alternative modes, which mutated Studio's bundled WordPress
installation. To rectify the situation, we remove the erroneous
wp-config.php file.

* test: Guard direct window references

Allow running tests in a Node.js test environment.

* test: Disable `ts-jest` diagnosis of `vendor/wp-now`

The TypeScript configuration is not properly configured to diagnose the
`vendor/wp-now` directory. As a result, any attempt to import or mock a
module within `vendor/wp-now` resulted in TypeScript compile errors.
Disabling test diagnostics for this directory mirrors the approach for
the application build.

* test: Assert `SiteServer` Playground mode

* refactor: Relocate `purgeWpConfig` to `wp-versions` module

The function relates to WordPress version management.

* refactor: Fix typo

* fix: Throw custom error rather than enforcing `wordpress` mode

An error will be thrown in this context regardless. Rather than
enforcing `wordpress` mode for a context that will not succeed, we throw
a unique, identifiable error that allows us to identify and track this
specific context in our error reporting tooling.

* refactor: Limit the capability of the `purgeWpConfig` method

Reduce the likelihood this purging might occur in an unintended
location by lowering the location logic into the method itself.