fix: Early return for non-empty, non-WordPress site directory
Basedd on the surrounding code, the lack of a return appears to be an
oversight. Introducing the return means creating a site in a non-empty,
non-WordPress directory no longer displays an error, and instead
dismisses the modal after moving the directory's content to the OS
trash. The apparent rationale for doing so is allowing the user to
retry with a clean directory.

A future improvement would be communicating the non-empty state to the
user, allowing them to take whatever action they deem appropriate.