fix: Improve site creation error handling (#4)
* fix: Early return for non-empty, non-WordPress site directory

Basedd on the surrounding code, the lack of a return appears to be an
oversight. Introducing the return means creating a site in a non-empty,
non-WordPress directory no longer displays an error, and instead
dismisses the modal after moving the directory's content to the OS
trash. The apparent rationale for doing so is allowing the user to
retry with a clean directory.

A future improvement would be communicating the non-empty state to the
user, allowing them to take whatever action they deem appropriate.

* fix: Correct error icon size

The flex layout caused the element to shrink.

* feat: Display a user-friendly error message

The error messages provided from the remote process can be highly
technical and unhelpful.

To start, they are all prefixed with confusing technical details:

> Error invoking remote method <name>

Also, the error origin can be very technical itself, e.g.:

> Error: ENOENT: no such file or directory, scandir '/Users/ajayghaghretiya/Library/Application Support/Studio/server-files/wordpress-versions/latest/wp-content/themes/twentytwentythree/parts'

To improve the user experience, we provide a generic, user-friendly
message instead, regardless of the origin.

* fix: Notify users of site creation failure due to invalid local path

Site creation silently failing is confusing. Instead, we should
communicate the error and ask that they retry.

* fix: Report failed working directory prep

Ensure we can monitor escalation of this type of error.

* fix: Notify users of site creation failure due to directory creation

Site creation silently failing is confusing. Instead, we should
communicate the error and ask that they retry.

* refactor: Remove unnecessary bespoke error message

This is no longer presented to the end-user. We can instead rely upon
the original error message.

* refactor: Rename error argument for consistency

Surrounding logic does not rely upon the `err` abbreviation.

* refactor: Hoist site creation error handling to renderer

Hoisting the `try`/`catch` to the main process invocation site ensures
that all errors originating from logic within the main process method
are caught. Previously, only certain portions of the logic were wrapped
in `try`/`catch`, meaning some errors would go unreported.

* fix: Site creation resets to proposed path when path is reset to default

Allow selecting the default site path (e.g., `~/Studio` on macOS) as a
means of resetting prior path selections. This allows users to recover
from mistakenly changing the local path to a non-empty, non-WordPress
directory.

* test: Assert add site error message