Fix filename used in `cp` during CI distribution
I've been using this automation in another app and noticed that the
filename to be copied was incorrect.

When working on DRYing the implementation, I didn't notice that the
existing code duplicated the logic to assemble the name using root and
suffix, etc. Having extracted that logic and its result in the
`:filename` property, the computed filename for `cp` had unnecessary
repetitions.