test: Improve stability of onboarding interactions
Combine `toBeVisible`'s interval with the ability to assert the precence
of one of multiple elements to ensure the UI is ready before
conditionally proceeding with the onboarding steps. This reduces the
likelihood that the UI is not ready when the test interactions begin and
also allows the Sites tests to succeed regardless of whether sites
already exists when the tests begin to run.