test: Delete site E2E case supports native dialogs (#186)
* test: Delete site E2E case supports native dialogs

Playwright lacks support for interacting with native dialogs, so we mock
the dialog module to simulate the user clicking the "Delete site"
confirmation button with "Delete site files from my computer" checked.

See: https://github.com/microsoft/playwright/issues/2143
See: https://github.com/microsoft/playwright/issues/8278#issuecomment-1009957411

* test: Update Onboarding continue button label text

* refactor: Fix typo

* test: Improve stability of onboarding interactions

Combine `toBeVisible`'s interval with the ability to assert the precence
of one of multiple elements to ensure the UI is ready before
conditionally proceeding with the onboarding steps. This reduces the
likelihood that the UI is not ready when the test interactions begin and
also allows the Sites tests to succeed regardless of whether sites
already exists when the tests begin to run.