feat: Rephrase download optimized version cancel option
This option is not so much a cancellation, but either a deferment of a
highly-recommended action or an acceptance of actively choosing to not
heed direct advice for solving an system issue. Rephrasing the option to
align with this better communicates the choice and makes the permanent
dismissal option more understandable.