Move back Lint to default queue
The caching mechanism is still broken, but at least we won't spam the
mac queue, which is more expensive and should be used to run steps that
legitimately need macOS.

However, happy to revert this in the interest to optimize CI speed.

The main reason for this is to have a working example of the changes in
the plugin.