Use `env` node for all `IS_DEV_BUILD` CI assignments
In https://github.com/Automattic/studio/pull/998#discussion_r1978739613
I got confirmation that it's okay to use this approach for the env var
setting.

Given it's already in use, because it would not work otherwise, in the
Windows steps, this commit updates all the other steps so we have a
consistent setup.